pyproject.toml
[project]
name = "pits-n-giggles"
description = "F1 Sim Racing Telemetry Analysis"
authors = [
    {name = "Ashwin Natarajan",email = "ashwin.nat@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "jinja2 (>=3.1.6,<4.0.0)",
    "iniconfig (>=2.1.0,<3.0.0)",
    "pyinstaller (>=6.12.0,<7.0.0)",
    "python-socketio (>=5.12.1,<6.0.0)",
    "quart (>=0.20.0,<0.21.0)",
    "gevent (>=24.11.1,<25.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "packaging (>=25.0,<26.0)",
    "msgpack (>=1.1.1,<2.0.0)",
    "pyzmq (>=27.0.0,<28.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "pyside6 (>=6.10.0,<7.0.0)",
    "markdown (>=3.10,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[tool.poetry.group.dev.dependencies]
tqdm = "^4.67.1"
deepdiff = "^8.4.2"
yappi = "^1.6.10"
snakeviz = "^2.2.2"
gdown = "^5.2.0"
colorama = "^0.4.6"
tabulate = "^0.9.0"
pylint = "3.2.6"
coverage = "^7.11.0"

